requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.24.3

# MongoDB
pymongo==4.6.1
dnspython==2.4.2

# Spark
pyspark==3.5.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Machine Learning
scikit-learn==1.3.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1

